pre-fetch notebooks; put notebook datastore in a separate file. todo: better typescript organization (first attempt at importing d.ts failed)
