+++ web/packages/web-client/package.json
@@ -42,7 +42,6 @@
     "js-cookie": "^2.2.0",
     "jwt-decode": "^2.2.0",
     "next": "^8.0.0-canary.18",
-    "nookies": "^2.0.1",
     "normalize.css": "^8.0.1",
     "react": "^16.8.1",
     "react-dom": "^16.8.1"
