+++ web/packages/web-client/pages/notebook.tsx
@@ -1,245 +1,44 @@
 import { PureComponent } from 'react';
-import auth from '../libs/auth';
-import fetch from 'isomorphic-unfetch';
-import getConfig from 'next/config';
+
+import {
+  Notebook,
+  startRequest,
+  callbackFunctions,
+  startListener,
+  createNotebook,
+  removeItemFromNotebook,
+  deleteNotebook,
+  notebookType,
+  notebooksType,
+  kubeState
+} from '../components/Notebook/datastore';
 
 import '../styles/pages/notebook.scss';
 
-const isServer = typeof window === 'undefined';
-
-// TODO: If user isn't logged in, and make it to this page
-// TODO: Enumerate waitingStatus 'reason's
-// after 401, tell them they need to log in
-const cfg = getConfig().publicRuntimeConfig.NOTEBOOK;
-const URL = cfg.URL;
-
-declare type runningStatus = {
-  started_at: number;
-};
-
-declare type waitingStatus = {
-  reason: string;
-};
-
-declare type terminatedStatus = {
-  exit_code: number;
-  finished_at: number;
-  started_at: number;
-  reason: string;
-};
-
-declare type notebook = {
-  svc_name: string;
-  name: string;
-  pod_name: string;
-  token: string;
-  creation_date: string;
-  svc_status: string;
-  pod_status: string; // The kubernetes status.phase val for pod
-  container_status: {
-    running: runningStatus | null;
-    waiting: waitingStatus | null;
-    terminated: terminatedStatus | null;
-  } | null;
-  condition: {
-    message: string | null;
-    reason: string | null;
-    status: string | null;
-    type: string | null;
-  } | null;
-};
-
-declare type notebooks = { [pod_name: string]: notebook };
-
 declare type props = {
   pageProps: {
-    notebooks: notebooks;
+    notebooks: notebookType;
   };
 };
 
 declare type state = {
   unauthorized: boolean;
-  notebooks: notebooks;
-  alive: notebook[];
+  notebooks: notebooksType;
+  alive: notebookType[];
   loading: number; //-1, 0, 1: failed, not, loading
 };
 
-enum kubeState {
-  Deleted = 'Deleted',
-  Running = 'Running'
-}
-
-enum kubeEvent {
-  ADDED = 'ADDDED',
-  DELETED = 'DELETED',
-  MODIFIED = 'MODIFIED'
-}
-
-declare type kubeWebsocketUpdate = {
-  event: kubeEvent;
-  resource: notebook;
-};
-
-const getAlive = (nbs: notebooks) =>
-  Object.values(nbs).filter((d: notebook) => d.svc_status != kubeState.Deleted);
-
-let resultsPromise: Promise<{ notebooks: notebooks; alive: notebook[] }>;
-
-let notebooks: notebooks = {};
-let alive: notebook[] = [];
-let initialized = false;
-
-const startRequest = () => {
-  if (resultsPromise) {
-    return resultsPromise;
-  }
-
-  resultsPromise = fetch(`${URL}/api`, {
-    headers: {
-      Authorization: `Bearer ${auth.accessToken}`
-    }
-  })
-    .then(d => d.json())
-    .then((aNotebooks: notebook[]) => {
-      const data: notebooks = {};
-
-      aNotebooks.forEach(n => {
-        data[n.pod_name] = n;
-      });
-
-      notebooks = data;
-      alive = getAlive(data);
-
-      return { notebooks, alive };
-    })
-    .finally(() => {
-      initialized = true;
-    });
-
-  return resultsPromise;
-};
-
-let isListening: boolean;
-const callbackFunctions: ((nb: notebooks, alive: notebook[]) => void)[] = [];
-
-const startListener = () => {
-  if (isListening) {
-    return;
-  }
-
-  const name = URL.replace(/http/, 'ws'); //automatically wss if https
-  const ws = new WebSocket(`${name}/api/ws`);
-
-  isListening = true;
-
-  ws.onmessage = ev => {
-    const data: kubeWebsocketUpdate = JSON.parse(ev.data);
-
-    const event = data.event;
-    const updated = data.resource;
-
-    const isDeleted =
-      updated.svc_status === kubeState.Deleted ||
-      updated.pod_status === kubeState.Deleted;
-
-    if (!(event === kubeEvent.DELETED || isDeleted)) {
-      notebooks[updated.pod_name] = updated;
-      alive = getAlive(notebooks);
-
-      // TODO: add errors
-      callbackFunctions.forEach(cb => {
-        cb(notebooks, alive);
-      });
-
-      return;
-    }
-
-    const existing = !!notebooks[updated.pod_name];
-
-    if (!existing) {
-      return;
-    }
-
-    callbackFunctions.forEach(cb => {
-      cb(notebooks, alive);
-    });
-
-    delete notebooks[updated.pod_name];
-    alive = getAlive(notebooks);
-
-    // TODO: add errors
-    callbackFunctions.forEach(cb => {
-      cb(notebooks, alive);
-    });
-  };
-};
-
-const createNotebook = async () => {
-  return fetch(`${URL}/api`, {
-    headers: {
-      Authorization: `Bearer ${auth.accessToken}`
-    },
-    method: 'POST'
-  })
-    .then(d => d.json())
-    .then((notebook: notebook) => {
-      notebooks[notebook.pod_name] = notebook;
-      alive = getAlive(notebooks);
-
-      return { notebooks, alive };
-    });
-};
-
-const removeItemFromNotebook = (notebook: notebook) => {
-  delete notebooks[notebook.pod_name];
-  alive = getAlive(notebooks);
-};
-
-const deleteNotebook = (notebook: notebook, skipRemove: boolean) => {
-  return fetch(`${URL}/api/${notebook.svc_name}`, {
-    headers: {
-      Authorization: `Bearer ${auth.accessToken}`
-    },
-    method: 'DELETE'
-  }).then(response => {
-    if (!response.ok && response.status !== 404) {
-      console.error(response);
-      throw Error(`${response.status}`);
-    }
-
-    if (!skipRemove) {
-      delete notebooks[notebook.pod_name];
-      alive = getAlive(notebooks);
-    }
-
-    return { notebooks, alive };
-  });
-};
-
-// Not used for now, until I get stable date from Kubernetes/notebook-api
-// (i.e not dependent on Date.now)
-// const getFormattedDate = (ms: number) => {
-//   const date = new Date(ms);
-//   let hours = date.getHours();
-
-//   if (hours > 12) {
-//     hours = 24 - hours;
-//   }
-
-//   const pm = hours >= 12 ? 'pm' : 'am';
-
-//   return `${date.getMonth() +
-//     1}/${date.getDate()}/${date.getFullYear()} at ${hours}:${date.getMinutes()}:${date.getSeconds()} ${pm}`;
-// };
+const isServer = typeof window === 'undefined';
 
 let loadingTimeout: NodeJS.Timeout | null;
+
 // TODO: decide whether we want to show only notebooks whose svc and pod status
 // TODO: check that there are no side-effects for mutating this.state.notebooks
 // are both Running
 // Argument against this is to give fine-grained insight into what Kube is doing
 // Because Kube is not a good queue, and will give out-of-order events
 // which may be easier for the user to understand, that for us to present always as in-order
-class Notebook extends PureComponent<any, state> {
+class NotebookPage extends PureComponent<any, state> {
   state: state = {
     loading: 0,
     unauthorized: false,
@@ -250,20 +49,20 @@ class Notebook extends PureComponent<any, state> {
   constructor(props: props) {
     super(props);
 
-    if (initialized === false) {
+    if (Notebook.initialized === false) {
       this.state.loading = 1;
 
       if (!isServer) {
         startRequest();
       }
     } else {
-      this.state.notebooks = Object.assign({}, notebooks);
-      this.state.alive = alive.slice(0);
+      this.state.notebooks = Object.assign({}, Notebook.notebooks);
+      this.state.alive = Notebook.aliveNotebooks.slice(0);
     }
   }
 
   componentDidMount() {
-    if (initialized) {
+    if (Notebook.initialized) {
       loadingTimeout = setTimeout(() => {
         this.setState({
           loading: 1
@@ -318,13 +117,13 @@ class Notebook extends PureComponent<any, state> {
   // https://jsperf.com/splice-vs-filter
   updateState = (success: boolean) => {
     this.setState({
-      notebooks: Object.assign({}, notebooks),
-      alive: alive.slice(0),
+      notebooks: Object.assign({}, Notebook.notebooks),
+      alive: Notebook.aliveNotebooks.slice(0),
       loading: success ? 0 : -1
     });
   };
 
-  deleteNotebook = (notebook: notebook) => {
+  deleteNotebook = (notebook: notebookType) => {
     // Do the deletion opportunisticly
     removeItemFromNotebook(notebook);
 
@@ -354,14 +153,12 @@ class Notebook extends PureComponent<any, state> {
               cursor: 'pointer'
             }}
           >
-            {this.state.alive.map((d: notebook) => (
+            {this.state.alive.map((d: notebookType) => (
               // Kubernetes is far too noisy, this prevents us from seeing pods that
               // are unreachable (since back-end/service is missing)
-              <span key={d.pod_name} className={'animated fadeIn fastest nb'}>
+              <span key={d.pod_name} className={'nb'}>
                 {d.svc_status === kubeState.Running &&
                 d.pod_status === kubeState.Running &&
-                // d.container_status &&
-                // d.container_status.running &&
                 d.condition &&
                 d.condition.status === 'True' &&
                 d.condition.type === 'Ready' ? (
@@ -393,33 +190,9 @@ class Notebook extends PureComponent<any, state> {
                   <span className="small">
                     Pod: <b>{d.pod_status}</b>
                   </span>
-                  {/* <span className="small">
-                    Service Name: <b>{d.svc_name}</b>
-                  </span> */}
                   <span className="small">
                     Pod Name: <b>{d.pod_name}</b>
                   </span>
-                  {/* <span className="small">
-                    Service: <b>{d.svc_status}</b>
-                  </span> */}
-
-                  {/* <span className="small">
-                    Container:{' '}
-                    {d.container_status === null ? (
-                      <b>Initializing</b>
-                    ) : d.container_status.waiting ? (
-                      <b>
-                        Waiting{' '}
-                        {d.container_status.waiting.reason
-                          ? '(' + d.container_status.waiting.reason + '_'
-                          : null}
-                      </b>
-                    ) : d.container_status.running ? (
-                      <b>Running</b>
-                    ) : (
-                      <b>Terminated</b>
-                    )}
-                  </span> */}
                   <span className="small">Created on: {d.creation_date}</span>
                 </a>
                 <i
@@ -438,4 +211,4 @@ class Notebook extends PureComponent<any, state> {
   }
 }
 
-export default Notebook;
+export default NotebookPage;
