+++ web/packages/web-client/styles/main.scss
@@ -120,7 +120,4 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
   table {
     border-collapse: collapse;
   }
-  .animated.fastest {
-    animation-duration: 100ms;
-  }
 }
